backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
langchain==0.1.4
langgraph==0.0.20
langchain-google-genai==0.0.11
langchain-community==0.0.16
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
chromadb==0.4.22
sentence-transformers==2.3.1
websockets==12.0
python-multipart==0.0.6
google-generativeai==0.3.2


